se crean servicios y controladores

+++ StructureSFYN/src/main/java/com/example/StructureSFYN/repositories/EmpresaRepository.java
@@ -1,8 +1,11 @@
 package com.example.StructureSFYN.repositories;
 
+import com.example.StructureSFYN.entities.Empresa;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.RepositoryDefinition;
+import org.springframework.stereotype.Repository;
+
 @Repository
-public interface EmpresaRepository extends JpaRepository <Empresa, Long> {
+public interface EmpresaRepository extends JpaRepository <Empresa, Integer> {
 
 }
